// packages/relay/src/lib/types/mirrorNode.ts
export interface LatestBlockNumberTimestamp {
  blockNumber: string | null;
  timeStampTo: string;
}